requirements.txt
# Core HTTP & Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
newspaper3k>=0.2.8
aiohttp>=3.9.0

# YouTube
yt-dlp>=2024.1.0
youtube-transcript-api>=0.6.2

# PubMed
biopython>=1.83

# NLP & Language
langdetect>=1.0.9
rake-nltk>=1.0.6
nltk>=3.8.0

# Web Dashboard
flask>=3.0.0
flask-cors>=4.0.0

# Data & Utilities
python-dateutil>=2.8.2
tqdm>=4.66.0
colorama>=0.4.6

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0

# Optional ML-based tagging (comment out if slow install)
# keybert>=0.7.0
# sentence-transformers>=2.2.0
